Arbitrary code execution fix
Thanks to @martia86 for pointing out the serious security flaw with arbitrary code execution. Before, there was no check on the key-input, which allowed users to create pastes with a .php extension, and then run any PHP code they wanted. As a result, I added a check to ensure the key does not contain the substring "php," thus (hopefully) fixing anymore possible exploits. 

I'd like to make it clear that this code is not intended to be used in production. This is something I made in about ten minutes and put on my GitHub, I don't expect any personal website or organization to be running this code. This is just the barebone, back-end implementation of text-hosting and file-deletion.

Thanks again.